Rename de comment a DiscussionComment para evitar conflictos.
